# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunGradeBody": {
        "additionalProperties": false,
        "description": "One grade for one of the run's samples on one human criterion.",
        "properties": {
          "criterion": {
            "minLength": 1,
            "title": "Criterion",
            "type": "string"
          },
          "grade": {
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Grade",
            "type": "integer"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "sample_id": {
            "minLength": 1,
            "title": "Sample Id",
            "type": "string"
          }
        },
        "required": [
          "sample_id",
          "criterion",
          "grade"
        ],
        "title": "RunGradeBody",
        "type": "object"
      },
      "RunGradesBody": {
        "additionalProperties": false,
        "description": "A batch of grades. Partial submission is normal; each is upserted on its own row.",
        "properties": {
          "graded_by": {
            "default": "unknown",
            "minLength": 1,
            "title": "Graded By",
            "type": "string"
          },
          "grades": {
            "items": {
              "$ref": "#/components/schemas/RunGradeBody"
            },
            "title": "Grades",
            "type": "array"
          }
        },
        "title": "RunGradesBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FreeWeight",
    "version": "1.2.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runs/{run_id}/grades": {
      "post": {
        "description": "Record grades on the run's samples; composites, aggregates and evidence follow.\n\nUpserted per ``(sample, criterion)``, so a batch sent twice — a client retrying after a dropped\nconnection — replaces rather than duplicates.\n\nRaises:\n    RunNotFound: No run has this id.\n    RunNotGradeable: See :func:`run_grading_endpoint`.\n    ValidationError: A sample outside the run, a criterion that is not a human one, or a grade\n        outside the criterion's scale — and then no grade of the batch lands.",
        "operationId": "run_grades_endpoint_api_v1_runs__run_id__grades_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunGradesBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Grades Endpoint Api V1 Runs  Run Id  Grades Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grade a goal run's samples",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/grading": {
      "get": {
        "description": "The samples a person grades on a completed goal run's human criteria, blinded and shuffled.\n\nRaises:\n    RunNotFound: No run has this id.\n    RunNotGradeable: ``409``: not a completed goal run with a human criterion, or its goal's\n        rubric has changed since the run.",
        "operationId": "run_grading_endpoint_api_v1_runs__run_id__grading_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Grading Endpoint Api V1 Runs  Run Id  Grading Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A goal run's blinded grading view",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}